Face-Mask Detection Using Python.
This program is made by Python. Here Machine Learning has been implemented using open cv and sklearn. In Order to run the program just run all the tabs (try using jupyter notebook for this). CLICK ESC to terminate the window/program. 
Note: Try to test this in the light as in dark the program always shows mask whether you are wearing it or not.